fix(pr): move code to keel-titus-plugin
Move to keel-titus-plugin since the implementation is coupled to titus.
